[PP] Added modules
