--- price-map/apps/scraper/scraper/__init__.py
@@ -1,1 +1,0 @@
-""""""
