+++ price-map/apps/scraper/scraper/index.py
@@ -1,11 +1,2 @@
-# from './services/base' import BaseScrapingService
-# from base-scraping.service import BaseScrapingService
-# base = BaseScrapingService()
-# print(base)
-
-#works tvar
-# from test import testFunc
-# print(testFunc())
-
-from base_scraping_service import BaseScrapingService
+from services.base_scraping_service import BaseScrapingService
 print(4)
