--- price-map/apps/scraper/scraper/test.py
@@ -1,3 +1,0 @@
-def testFunc():
-  result = 5 + 6
-  return result
